# api/api/openapi.yaml
openapi: 3.0.3
info:
  title: Minori API
  description: OAS of Minori API
  version: 0.dev
paths:
...
  /minors/minor:
    get:
      tags:
        - minors
      description: Get all minors
      operationId: minorsMinor.get
      parameters:
        - schema:
            type: string
            default: ""
          name: name
          in: query
        - schema:
            type: string
            default: ""
          name: surname
          in: query
        - schema:
            type: string
            default: ""
          name: fiscal_code
          in: query
        - schema:
            type: string
            default: ""
          name: entry_date
          in: query
        - schema:
            type: integer
          name: page
          in: query
          required: true
        - schema:
            type: integer
          name: limit
          in: query
          required: true
        - schema:
            type: string
            default: ""
          in: query
          name: status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Minor"
    post:
      tags:
        - minors
      description: Permit to insert a new minor
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
              required:
                - name
                - surname
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
      operationId: minorsMinor.post
...
  /adults/adult:
    get:
      tags:
        - adults
      description: Get all adults
      operationId: adultsAdult.get
      parameters:
        - schema:
            type: string
            default: ""
          name: name
          in: query
        - schema:
            type: string
            default: ""
          name: surname
          in: query
        - schema:
            type: string
            default: ""
          name: fiscal_code
          in: query
        - schema:
            type: string
            default: ""
          name: entry_date
          in: query
        - schema:
            type: integer
          name: page
          in: query
          required: true
        - schema:
            type: integer
          name: limit
          in: query
          required: true
        - schema:
            type: string
            default: ""
          in: query
          name: status
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Adult"
    post:
      tags:
        - adults
      description: Permit to insert a new adult
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
              required:
                - name
                - surname
            examples: {}
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
      operationId: adultsAdult.post
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: api.middleware.authn.authn_bearer
    CookieAuth:
      name: access_token_cookie
      type: apiKey
      in: cookie
      x-apikeyInfoFunc: api.middleware.authn.authn_cookie
  schemas:
...
    Minor:
      type: object
      description: Describe a minor object
      properties:
        _id:
          type: object
          $ref: "#/components/schemas/ObjectId"
        name:
          type: string
        surname:
          type: string
      required:
        - _id
        - name
        - surname
...
    Adult:
      type: object
      description: Describe an adult object
      properties:
        _id:
          type: object
          $ref: "#/components/schemas/ObjectId"
        name:
          type: string
        surname:
          type: string
      required:
        - _id
        - name
        - surname
...
    ObjectId:
      type: string
      title: ObjectId
      format: uuid
      minLength: 24
      maxLength: 24
      pattern: "^[0-9a-f]+$"
      description: MongoDB ObjectId
...
    Success:
      type: object
      title: Success
      properties:
        success:
          type: boolean
      required:
        - success
...
  responses:
    SuccessResponse:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
...
security:
  - BearerAuth: []
  - CookieAuth: []
...